reattach

+++ autotrace/autotrace.py
@@ -861,15 +861,19 @@ def process_args():
 		time.sleep(1)
 	if args.commands == [] and not args.replayfile and not args.replay:
 		pid, command = get_last_run_pid()
-		if pid:
-			args.commands.append("""echo bg command being tracked is: """ + command)
-			args.commands.append('iostat 1')
-			if platform.system() != 'Darwin':
- 				args.commands.append("""bash -c 'while true; do cat /proc/""" + str(pid) + """/status; sleep 2; done""")
-			args.commands.append('vmstat 1')
-		else:
+		if pid is None and command is None:
 			print('No background process found on this terminal session.')
 			sys.exit(1)
+		if command:
+			args.commands.append(command)
+		elif pid is not None:
+			args.commands.append("""echo bg command being tracked is: """ + command)
+		args.commands.append('iostat 1')
+		if pid is not None and platform.system() != 'Darwin':
+ 			args.commands.append("""bash -c 'while true; do cat /proc/""" + str(pid) + """/status; sleep 2; done""")
+		elif pid is None and platform.system() != 'Darwin':
+ 			args.commands.append("""bash -c 'while true; do cat /proc/PID/status; sleep 2; done""")
+		args.commands.append('vmstat 1')
 	# Validate END
 	# BUG! if logtimestep is false it's broked - is it?
 	#args.logtimestep = True
@@ -882,17 +886,9 @@ def get_last_run_pid(encoding='utf-8'):
 	# The grep gets all processes with the same tty
 	pses = pexpect.run("""bash -c '(export LC_ALL=C; ps -a -o etime=,tt=,pid=,args= | sort -r)'""").decode(encoding).strip()
 	pses = pses.split('\r\n')
-	#print(mytty)
-	#print(pses)
 	pses_on_this_tty = []
 	for line in pses:
 		line_list = re.split(r'\s+', line.strip())
-		#print('===========')
-		#print(type(line_list[1]))
-		#print(type(mytty))
-		#print(line_list[1])
-		#print(mytty)
-		#print(line_list[1] == mytty)
 		if line_list[1] == mytty:
 			pses_on_this_tty.append(line_list)
 	if len(pses_on_this_tty) == 2:
@@ -909,30 +905,40 @@ def get_last_run_pid(encoding='utf-8'):
 		clear_screen()
 		print('''
 
- OK, you are about to attach to the origin command:
+OOK, you are about to attach to the origin command:
 
 	''' + command + '''
 
 which has been suspended until you choose to continue or quit.
 
 Output of the original command is not easily redirected away from this terminal.
-However, you can replay the output once the process is finished with:
+However, you can replay the autotrace once the process is finished with:
 
-	autotrace --replay <DIR>
+	autotrace --replay <LOGDIR>
 
-OK?
+You have some choices:
 
-''')
-		message = "Enter to continue, q to quit and restart the original command, z to just quit..."
+- Enter           - restart the command and trace
+- q and Enter     - restart the original command and quit
+- z and Enter     - change nothing, just quit
+- r and Enter     - kill off the already-started process and re-run with autotrace
+
+=> ''')
 		if PY3:
-			resp = input(message)
+			resp = input()
 		else:
-			resp = raw_input(message)
+			resp = raw_input()
 		# Redirect output? Can't without other deps, but can replay
-		pexpect.run('kill -CONT ' + pid)
+		if resp in ('z','Z'):
+			sys.exit(0)
+		if resp in ('r','R'):
+			pexpect.run('kill -KILL ' + pid)
+			return None, command
+		elif resp in ('q','Q','',''):
+			pexpect.run('kill -CONT ' + pid)
 		if resp in ('q','Q'):
-			sys.exit(1)
-		return int(pid), ' '.join(command)
+			sys.exit(0)
+		return int(pid), command
 	return None, None
